script/requirements.txt
requests==2.31.0
newspaper3k==0.2.8
beautifulsoup4==4.12.3
nltk==3.8.1
lxml==4.9.3
python-dotenv==1.0.0
schedule==1.2.2